Revert moving of ValidateCustomLess logic.  Allow for attribute specific setting validation rules.
